Got various addresses
These all seem to fit into some sort of table that's consistently in the same place in memory. If I had to guess, the kills, scores, etc. that belong to the current user are determined by some sort of ID the user gets assigned when they enter the server. This is probably the same ID that you get shown if you enter `status` in the console in CS:GO.